feat: Added new AbilityExt module that contains all data structures that support BP-only bespoke ability types (Effect/Kit/Trait)
