+++ Config/DefaultCkFoundation.ini
@@ -127,3 +127,25 @@
 +EnumRedirects=(OldName="/Script/CkNet.ECk_Net_NetModeType",NewName="/Script/CkEcs.ECk_Net_NetModeType")
 +EnumRedirects=(OldName="/Script/CkNet.ECk_Net_NetExecutionPolicy",NewName="/Script/CkEcs.ECk_Net_NetExecutionPolicy")
 +EnumRedirects=(OldName="/Script/CkNet.ECk_Net_EntityNetRole",NewName="/Script/CkEcs.ECk_Net_EntityNetRole")
+
++ClassRedirects=(OldName="/Script/CkAbility.Ck_Ability_TraitCondition_Interface",NewName="/Script/CkAbilityExt.Ck_Ability_TraitCondition_Interface")
++ClassRedirects=(OldName="/Script/CkAbility.Ck_Ability_TraitPayoff_Interface",NewName="/Script/CkAbilityExt.Ck_Ability_TraitPayoff_Interface")
++ClassRedirects=(OldName="/Script/CkAbility.Ck_Ability_Effect_Interface",NewName="/Script/CkAbilityExt.Ck_Ability_Effect_Interface")
+
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_ActivationTrigger_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_ActivationTrigger_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_EventsForwarding_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_EventsForwarding_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_Stacking_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_Stacking_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_CueStacking_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_CueStacking_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_StackDurationRefresh_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_StackDurationRefresh_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_StackPeriodReset_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_StackPeriodReset_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_StackOverflow_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_StackOverflow_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_StackLimit_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_StackLimit_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_StackExpiration_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_StackExpiration_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_Modifier_MagnitudeCalculation_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_Modifier_MagnitudeCalculation_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_ActivationRequirementsNotMet_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_ActivationRequirementsNotMet_Policy")
++EnumRedirects=(OldName="/Script/CkAbility.ECk_Ability_Duration_Policy",NewName="/Script/CkAbilityExt.ECk_Ability_Duration_Policy")
+
++StructRedirects=(OldName="/Script/CkAbility.Ck_Ability_Modifier_Magnitude",NewName="/Script/CkAbilityExt.Ck_Ability_Modifier_Magnitude")
++StructRedirects=(OldName="/Script/CkAbility.Ck_Ability_Modifier",NewName="/Script/CkAbilityExt.Ck_Ability_Modifier")
++StructRedirects=(OldName="/Script/CkAbility.Ck_Ability_TraitPayoff_Instances",NewName="/Script/CkAbilityExt.Ck_Ability_TraitPayoff_Instances")
++StructRedirects=(OldName="/Script/CkAbility.Ck_Ability_TraitCondition_Instances",NewName="/Script/CkAbilityExt.Ck_Ability_TraitCondition_Instances")
