+++ Source/CkAbilityExt/CkAbilityExt.cpp
@@ -1,0 +1,15 @@
+#include "CkAbilityExt.h"
+
+#define LOCTEXT_NAMESPACE "FCkAbilityExtModule"
+
+void FCkAbilityExtModule::StartupModule()
+{
+}
+
+void FCkAbilityExtModule::ShutdownModule()
+{
+}
+
+#undef LOCTEXT_NAMESPACE
+
+IMPLEMENT_MODULE(FCkAbilityExtModule, CkAbilityExt)
