+++ Source/CkAbilityExt/CkAbilityExt_Log.cpp
@@ -1,0 +1,12 @@
+#include "CkAbilityExt_Log.h"
+
+// --------------------------------------------------------------------------------------------------------------------
+
+DEFINE_LOG_CATEGORY(CkAbilityExt);
+
+namespace ck::ability_extension
+{
+	CK_REGISTER_LOG_FUNCTIONS(CkAbilityExt);
+}
+
+// --------------------------------------------------------------------------------------------------------------------
