+++ Source/CkAbilityExt/CkAbilityExt_Log.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "CkCore/Log/CkLog.h"
+
+// --------------------------------------------------------------------------------------------------------------------
+
+CKABILITYEXT_API DECLARE_LOG_CATEGORY_EXTERN(CkAbilityExt, Log, All);
+
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace ck::ability_extension
+{
+    CK_DEFINE_LOG_FUNCTIONS(CkAbilityExt);
+}
+
+// --------------------------------------------------------------------------------------------------------------------
