+++ Source/CkAbilityExt/Public/CkAbilityExt/Ability/CkAbilityExt_Fragment_Data.cpp
@@ -1,0 +1,3 @@
+#include "CkAbilityExt_Fragment_Data.h"
+
+// --------------------------------------------------------------------------------------------------------------------
